# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Druglane Cloud API",
        "description": "Cloud-based management system for offline desktop inventory management applications. This API handles license management, backup storage, and provides online access to inventory data.",
        "contact": {
            "email": "admin@druglane.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/license/validate": {
            "post": {
                "tags": [
                    "License"
                ],
                "summary": "Validate license key",
                "description": "Check if a license key is valid and active. Used by desktop applications.",
                "operationId": "8248842c5ebfedab29a0fdf2f6b4a7aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "license_key"
                                ],
                                "properties": {
                                    "license_key": {
                                        "type": "string",
                                        "example": "ABCD-EFGH-IJKL-MNOP-QRST-UVWX-YZ12-3456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "License is valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License is valid"
                                        },
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Acme Corporation"
                                                },
                                                "is_activated": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "license_status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "license_expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "License expired or suspended",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License has expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid license key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid license key"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/license/activate": {
            "post": {
                "tags": [
                    "License"
                ],
                "summary": "Activate license",
                "description": "Activate a license on a specific machine. One-time activation only. Used by desktop applications.",
                "operationId": "118dd810beed48f8e34333705e597088",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "license_key",
                                    "machine_id"
                                ],
                                "properties": {
                                    "license_key": {
                                        "type": "string",
                                        "example": "ABCD-EFGH-IJKL-MNOP-QRST-UVWX-YZ12-3456"
                                    },
                                    "machine_id": {
                                        "description": "Unique machine identifier",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "DESKTOP-ABC123-UUID-XYZ789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "License activated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License activated successfully"
                                        },
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Acme Corporation"
                                                },
                                                "license_key": {
                                                    "type": "string"
                                                },
                                                "contact_email": {
                                                    "type": "string"
                                                },
                                                "activated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "license_expires_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "License cannot be activated (suspended or expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License has been suspended"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid license key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid license key"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "License already activated on another machine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License has already been activated on another machine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}